Add validators for new resource fields
Each new field needs a validator to check if the input is correct
or not. Since these fields are optional the validators also need
to ignore (approve) the input when the isn't any value. Because of
that we need to reimplement the positive integer validator from
ckan core which fails (raises an exception) when there is no value.