Check if year has value before passing it to the int_validator
If a None value is passed to the int_validator it fails as it
expects an Integer. Year is an optional field so None should be
allowed as a value. That means we we need to first check if the
year has any value and if so then we can move onto year validation
if not we don't do any validation.